requirements.txt
streamlit
pandas
pandasai
pandasai-litellm
litellm
python-dotenv
plotly
altair
vl-convert-python